Refactor groups controller and finalized refactor of backend
